Finished global data system

+++ src/cc/isotopestudio/DailyQuest/task/DailyUpdate.java
@@ -1,7 +1,9 @@
 package cc.isotopestudio.DailyQuest.task;
 
+import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
+import java.text.SimpleDateFormat;
 
 import org.bukkit.scheduler.BukkitRunnable;
 
@@ -17,42 +19,61 @@ public class DailyUpdate extends BukkitRunnable {
 
 	@Override
 	public void run() {
+		boolean ifTableExist = true;
 		Statement statement = null;
 		try {
 			statement = DailyQuest.c.createStatement();
 		} catch (SQLException e1) {
-			e1.printStackTrace();
 		}
 		// Check if a new day
-
-		// if true
-
-		// Delete Database
+		ResultSet res = null;
+		java.util.Date todayDate = new java.util.Date();
+		SimpleDateFormat format = new SimpleDateFormat("yyyyMMdd");
+		String today = format.format(todayDate);
+		System.out.println(today);
 		try {
-			statement.executeUpdate("drop table global;");
-			statement.executeUpdate("drop table globalstage3;");
-			statement.executeUpdate("drop table players;");
-		} catch (SQLException e) {
-			e.printStackTrace();
+			res = statement.executeQuery("select * from global;");
+		} catch (SQLException e1) {
+			ifTableExist = false;
 		}
-		System.out.println("Delete databse");
-		// Create table
+		if (ifTableExist)
+			try {
+				if (res.next()) {
+					ifTableExist = true;
+					java.util.Date dayDate = new java.util.Date((res.getDate("today").getTime()));
+					String day = format.format(dayDate);
+					System.out.println(day);
+					if (day.equals(today)) {
+						System.out.println("true");
+						return;
+					}
+				}
+			} catch (SQLException e1) {
+				e1.printStackTrace();
+			}
 		try {
+			if (ifTableExist) {
+				// Delete tables
+				statement.executeUpdate("drop table global;");
+				statement.executeUpdate("drop table globalstage3;");
+				statement.executeUpdate("drop table players;");
+				System.out.println("Deleted tables");
+			}
+			// Create tables
 			statement.executeUpdate("create table global(" + "today date not null primary key,"
 					+ "stage1Type tinytext not null," + "stage1Num tinyint not null," + "stage2Type tinytext not null,"
 					+ "stage2Num tinyint not null);");
-			statement.executeUpdate(
-					"create table globalstage3(" + " id int unsigned not null auto_increment primary key,"
-							+ " X tinyint not null," + " Y tinyint not null," + " Z tinyint not null);");
+			statement.executeUpdate("create table globalstage3("
+					+ " id int unsigned not null auto_increment primary key," + " world text not null,"
+					+ " X SMALLINT not null," + " Y SMALLINT not null," + " Z SMALLINT not null);");
 			statement.executeUpdate("create table players(" + " name char(20) not null primary key,"
 					+ " stage tinyint not null," + " step tinyint not null," + " time tinyint not null);");
-			statement.executeUpdate("insert into global values(20160316,\"ZOMBIE\",5,\"IRON_ORE\",2);");
-			statement.executeUpdate("insert into globalstage3 values(NULL,2,1,-50);");
-		} catch (SQLException e) {
-			e.printStackTrace();
+			System.out.println("Created tables");
+		} catch (SQLException e1) {
+			e1.printStackTrace();
 		}
 		// Update global information
-		GlobalData.update(plugin);
+		GlobalData.update(plugin, statement, today);
 
 	}
